% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run IAO Reader App}
\usage{
run_shiny_app(port = 8080)
}
\arguments{
\item{port}{a port on which the Shiny application will be hosted.}
}
\description{
This function is responsible for starting the core application
 of this package. The application itself is located inside the
 \code{iaoreader} package files.
}
